Move `KafkaClient` to `RD*` namespace
Motivation:

Our `KafkaClient` class has evolved into a wrapper of `rd_kafka_*` C
functions. Therefore I thought it is reasonable to move it to the
`RDKafka*` namespace (conceptually) and remove the imports of `CRDKafka`
in `KafkaConsumer` and `KafkaProducer`.

Modifications:

* `KafkaConsumer`: remove `CRDKafka` import
* `KafkaProducer`: remove `CRDKafka` import
* rename `KafkaClient -> `RDKafkaClient` and move it to the `RDKafka`
  folder
* delete `RDKafka.swift` and move `create/makeClient` factory method to
  `RDKafkaClient`
