+++ Sources/SwiftKafka/RDKafka/RDKafkaTopicConfig.swift
@@ -35,7 +35,7 @@ struct RDKafkaTopicConfig {
     /// - Parameter value: The new value of the configuration property to be changed.
     /// - Throws: A ``KafkaError`` if setting the value failed.
     static func set(configPointer: OpaquePointer, key: String, value: String) throws {
-        let errorChars = UnsafeMutablePointer<CChar>.allocate(capacity: KafkaClient.stringSize)
+        let errorChars = UnsafeMutablePointer<CChar>.allocate(capacity: RDKafkaClient.stringSize)
         defer { errorChars.deallocate() }
 
         let configResult = rd_kafka_topic_conf_set(
@@ -43,7 +43,7 @@ struct RDKafkaTopicConfig {
             key,
             value,
             errorChars,
-            KafkaClient.stringSize
+            RDKafkaClient.stringSize
         )
 
         if configResult != RD_KAFKA_CONF_OK {
