+++ Tests/IntegrationTests/SwiftKafkaTests.swift
@@ -54,7 +54,7 @@ final class SwiftKafkaTests: XCTestCase {
         )
 
         // TODO: ok to block here? How to make setup async?
-        let client = try RDKafka.createClient(
+        let client = try RDKafkaClient.makeClient(
             type: .consumer,
             configDictionary: basicConfig.dictionary,
             events: [],
@@ -71,7 +71,7 @@ final class SwiftKafkaTests: XCTestCase {
         )
 
         // TODO: ok to block here? Problem: Tests may finish before topic is deleted
-        let client = try RDKafka.createClient(
+        let client = try RDKafkaClient.makeClient(
             type: .consumer,
             configDictionary: basicConfig.dictionary,
             events: [],
